refactor(users): extract clerk and navbar profile modules from users actions
Move deleteClerkUser and navbar profile fetching into dedicated modules, and drop "use server" from festival definitions/helpers so type-only imports stay client-safe.
